requirements.txt
Django==4.2.10
psycopg2-binary==2.9.9
Pillow==10.2.0
django-crispy-forms==2.1
crispy-tailwind==1.0.0
django-ckeditor==6.7.0
django-tailwind==3.8.0
django-compressor==4.4
python-decouple==3.8
gunicorn==21.2.0
whitenoise==6.6.0